Use `getFilteredCards` in overview
