--- frontend/public/locales/de-DE/complete-progress.json
@@ -1,3 +1,0 @@
-{
-  "youHave": "{{nCardsOwned}} / {{nTotalCards}} Karten"
-}
