+++ frontend/public/locales/de-DE/expansion-overview.json
@@ -1,5 +1,6 @@
 {
   "probabilityNewCard": "Chance auf eine neue Karte",
   "totalCards": "Anzahl der Karten",
-  "percentage": "Prozent"
+  "percentage": "Prozent",
+  "youHave": "{{nCardsOwned}} / {{nTotalCards}} Karten"
 }
