--- frontend/public/locales/en-US/complete-progress.json
@@ -1,3 +1,0 @@
-{
-  "youHave": "{{nCardsOwned}} / {{nTotalCards}} cards"
-}
