--- frontend/public/locales/es-ES/complete-progress.json
@@ -1,3 +1,0 @@
-{
-  "youHave": "Tienes {{nCardsOwned}} de {{nTotalCards}} cartas"
-}
