--- frontend/public/locales/fr-FR/complete-progress.json
@@ -1,3 +1,0 @@
-{
-  "youHave": "Vous possédez {{nCardsOwned}} cartes sur {{nTotalCards}}"
-}
