+++ frontend/public/locales/fr-FR/expansion-overview.json
@@ -1,5 +1,6 @@
 {
   "probabilityNewCard": "Probabilité d'avoir une nouvelle carte par paquet",
   "totalCards": "Nombre de cartes total",
-  "percentage": "pourcentage"
+  "percentage": "pourcentage",
+  "youHave": "Vous possédez {{nCardsOwned}} cartes sur {{nTotalCards}}"
 }
