--- frontend/public/locales/it-IT/complete-progress.json
@@ -1,3 +1,0 @@
-{
-  "youHave": "Hai {{nCardsOwned}} carte su {{nTotalCards}}"
-}
