+++ frontend/public/locales/it-IT/expansion-overview.json
@@ -1,5 +1,6 @@
 {
   "probabilityNewCard": "Probabilità di ottenere nuove carte",
   "totalCards": "Carte totali",
-  "percentage": "percentage"
+  "percentage": "percentage",
+  "youHave": "Hai {{nCardsOwned}} carte su {{nTotalCards}}"
 }
