--- frontend/public/locales/pt-BR/complete-progress.json
@@ -1,3 +1,0 @@
-{
-  "youHave": "Você tem {{nCardsOwned}} de {{nTotalCards}} cartas únicas"
-}
