+++ frontend/src/components/Mission.tsx
@@ -5,7 +5,8 @@ import { useTranslation } from 'react-i18next'
 import { Tooltip } from 'react-tooltip'
 import FancyCard from '@/components/FancyCard.tsx'
 import useWindowDimensions from '@/hooks/useWindowDimensionsHook.ts'
-import { getCardById, getCardByInternalId, pullRateForSpecificMission } from '@/lib/CardsDB.ts'
+import { getCardById, getCardByInternalId } from '@/lib/CardsDB.ts'
+import { pullRateForSpecificMission } from '@/lib/stats'
 import { getCardNameByLang } from '@/lib/utils.ts'
 import { useCollection, useSelectedCard } from '@/services/collection/useCollection'
 import type { CollectionRow, Mission as MissionType } from '@/types'
