+++ frontend/src/pages/collection/CardDetail.tsx
@@ -9,7 +9,8 @@ import { Button } from '@/components/ui/button'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Radio, RadioIndicator, RadioItem } from '@/components/ui/radio'
 import { Sheet, SheetContent, SheetHeader, SheetTitle } from '@/components/ui/sheet'
-import { craftingCost, getCardByInternalId, getExpansionById, pullRateForSpecificCard } from '@/lib/CardsDB.ts'
+import { craftingCost, getCardByInternalId, getExpansionById } from '@/lib/CardsDB.ts'
+import { pullRateForSpecificCard } from '@/lib/stats'
 import { getCardNameByLang } from '@/lib/utils'
 import { useCollection, useDeleteCard, useSelectedCard } from '@/services/collection/useCollection'
 import type { Card, CollectionRow } from '@/types'
