+++ frontend/src/pages/collection/MissionDetail.tsx
@@ -2,7 +2,8 @@ import i18n from 'i18next'
 import { useTranslation } from 'react-i18next'
 import FancyCard from '@/components/FancyCard.tsx'
 import { Sheet, SheetContent, SheetHeader, SheetTitle } from '@/components/ui/sheet'
-import { getCardById, getExpansionById, pullRateForSpecificCard } from '@/lib/CardsDB.ts'
+import { getCardById, getExpansionById } from '@/lib/CardsDB.ts'
+import { pullRateForSpecificCard } from '@/lib/stats'
 import { getCardNameByLang } from '@/lib/utils.ts'
 import type { ExpansionId } from '@/types'
