+++ frontend/src/pages/overview/components/ExpansionOverview.tsx
@@ -3,22 +3,20 @@ import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useMediaQuery } from 'react-responsive'
 import { BarChartComponent } from '@/components/BarChart.tsx'
-import * as CardsDB from '@/lib/CardsDB.ts'
-import { CompleteProgress } from '@/pages/overview/components/CompleteProgress.tsx'
+import { pullRate } from '@/lib/stats'
 import { GradientCard } from '@/pages/overview/components/GradientCard.tsx'
-import { useCollection } from '@/services/collection/useCollection'
-import type { CollectionRow, Expansion, Rarity } from '@/types'
+import type { Card, Expansion } from '@/types'
 import { Carousel } from './Carousel'
+import { CompleteProgress } from './CompleteProgress'
 
 interface ExpansionOverviewProps {
   expansion: Expansion
-  rarityFilter: Rarity[]
-  numberFilter: number
+  collectedCards: Card[]
+  wantedCards: Card[]
+  availableCards: Card[]
   deckbuildingMode: boolean
 }
-export function ExpansionOverview({ expansion, rarityFilter, numberFilter, deckbuildingMode }: ExpansionOverviewProps) {
-  const { data: ownedCards = new Map<number, CollectionRow>() } = useCollection()
-
+export function ExpansionOverview({ expansion, collectedCards, availableCards, wantedCards, deckbuildingMode }: ExpansionOverviewProps) {
   const { t } = useTranslation(['expansion-overview', 'common/sets', 'common/packs'])
 
   const isMobile = useMediaQuery({ query: '(max-width: 767px)' })
@@ -31,7 +29,7 @@ export function ExpansionOverview({ expansion, rarityFilter, numberFilter, deckb
     }
     const chartData = packs.map((pack) => ({
       packName: t(pack.name, { ns: 'common/packs' }),
-      percentage: CardsDB.pullRate({ ownedCards, expansion, pack, rarityFilter, numberFilter, deckbuildingMode }),
+      percentage: pullRate(wantedCards, expansion, pack, deckbuildingMode),
       fill: pack.color,
     }))
 
@@ -41,7 +39,7 @@ export function ExpansionOverview({ expansion, rarityFilter, numberFilter, deckb
       highestProbabilityPack,
       chartData,
     }
-  }, [ownedCards, expansion, rarityFilter, numberFilter, deckbuildingMode])
+  }, [wantedCards, expansion, deckbuildingMode])
 
   return (
     <>
@@ -75,24 +73,16 @@ export function ExpansionOverview({ expansion, rarityFilter, numberFilter, deckb
               </>
             )}
             <div className="col-span-8 snap-start flex-shrink-0 w-full border-1 border-neutral-700 border-solid rounded-lg p-4 sm:p-8">
-              <CompleteProgress
-                title={t('totalCards')}
-                expansion={expansion}
-                rarityFilter={rarityFilter}
-                numberFilter={numberFilter}
-                deckbuildingMode={deckbuildingMode}
-              />
+              <CompleteProgress title={t('totalCards')} collected={collectedCards.length} available={availableCards.length} />
               {expansion.packs.length > 1 &&
                 expansion.packs.map((pack) => (
                   <CompleteProgress
                     key={pack.name}
-                    rarityFilter={rarityFilter}
+                    className="mt-4"
                     title={t(pack.name, { ns: 'common/packs' })}
-                    expansion={expansion}
-                    packName={pack.name}
-                    numberFilter={numberFilter}
-                    deckbuildingMode={deckbuildingMode}
                     barColor={pack.color}
+                    collected={collectedCards.filter((c) => c.pack === pack.name).length}
+                    available={availableCards.filter((c) => c.pack === pack.name).length}
                   />
                 ))}
             </div>
@@ -114,24 +104,16 @@ export function ExpansionOverview({ expansion, rarityFilter, numberFilter, deckb
             </>
           )}
           <div className="col-span-4 lg:col-span-2">
-            <CompleteProgress
-              title={t('totalCards')}
-              expansion={expansion}
-              rarityFilter={rarityFilter}
-              numberFilter={numberFilter}
-              deckbuildingMode={deckbuildingMode}
-            />
+            <CompleteProgress title={t('totalCards')} collected={collectedCards.length} available={availableCards.length} />
             {expansion.packs.length > 1 &&
               expansion.packs.map((pack) => (
                 <CompleteProgress
                   key={pack.name}
-                  rarityFilter={rarityFilter}
-                  numberFilter={numberFilter}
+                  className="mt-4"
                   title={t(pack.name, { ns: 'common/packs' })}
-                  expansion={expansion}
-                  packName={pack.name}
-                  deckbuildingMode={deckbuildingMode}
                   barColor={pack.color}
+                  collected={collectedCards.filter((c) => c.pack === pack.name).length}
+                  available={availableCards.filter((c) => c.pack === pack.name).length}
                 />
               ))}
           </div>
